requirements.txt
# Core deep learning
torch>=1.13.0
torchvision>=0.14.0

# HuggingFace diffusers ecosystem (main dependency for the refactored code)
diffusers>=0.20.0
transformers>=4.25.0
accelerate>=0.20.0

# Configuration
PyYAML>=6.0

# Data science utilities
numpy>=1.23.0
tqdm>=4.64.0

# Image processing
opencv-python>=4.6.0
Pillow>=9.0.0

# Logging and experiment tracking
tensorboardX>=2.5.0
wandb>=0.13.0

# Optional (only needed if running legacy scripts)
# lmdb>=1.3.0
# pandas>=1.5.0
# scipy>=1.9.0
